checking out works like a boss

+++ Scripts/ProceedToCheck.js
@@ -1,31 +1,87 @@
-const cartItems = JSON.parse(sessionStorage.getItem("cart"));
+
+// window.onhashchange = function() {
+//     //blah blah blah
+    
+// location.reload();
+//    }
+//window.location=document.referrer;
+
 let user = JSON.parse(localStorage.getItem("user"));
 const buyURL = "https://lamp.ms.wits.ac.za/~s1814731/MPphpfiles/MPBuy.php";
 const cartUrl = "https://lamp.ms.wits.ac.za/~s1814731/MPphpfiles/Cart/MPGetCart.php";
 const updateUserUrl = "https://lamp.ms.wits.ac.za/~s1814731/MPphpfiles/Cart/MPReturnUser.php";
 const updateCartUrl = "https://lamp.ms.wits.ac.za/~s1814731/MPphpfiles/Cart/MPUpdateCartItem.php";
 const DeleteCartItemUrl = "https://lamp.ms.wits.ac.za/~s1814731/MPphpfiles/Cart/MPDeleteCartItem.php";
+let cartItems;
 
-console.log(cartItems);
+//getCartItems();
+setVariables();
+
+//console.log( user);
+
+ function getCartItems(){
+   const promise = new Promise(resolve =>{
+     //console.log("in promise");   
+     $.getJSON(cartUrl , {userID : user.UserID} , function(results){
+        console.log("in promise2"); 
+    // sessionStorage.removeItem("cart");
+    // sessionStorage.setItem("cart" , JSON.stringify(results));
+    // cartItems = JSON.parse(sessionStorage.getItem("cart"));
+    resolve();
+    });
+    
+   });
+    
+   promise.then(results => {
+    //sessionStorage.removeItem("cart");
+    // sessionStorage.setItem("cart" , JSON.stringify(results));
+    // cartItems = JSON.parse(sessionStorage.getItem("cart"));
+    console.log("from item promise");
+   });
+
+}
+
+
+ 
 const ConfirmPurchase = document.getElementById("confirmPurchase"); //This the buy button
 let sumTotal = 0;
 
-//console.log(cartItems , user);
+function setVariables() {
+    
+    const promises = new Promise( resolve =>{  
+        $.getJSON(cartUrl , {userID : user.UserID} , function(results){
+            console.log("in promise3"); 
+        sessionStorage.removeItem("cart");
+        sessionStorage.setItem("cart" , JSON.stringify(results));
+        cartItems = JSON.parse(sessionStorage.getItem("cart"));
+        resolve();
+        });
+        //resolve(getCartItems());
+    });
+    
+    promises.then(()=>{
+      
+        cartItems = JSON.parse(sessionStorage.getItem("cart"));
+        console.log('Trying out promises : ', cartItems);
+        if (cartItems === null) return;
 
-let setVariables = function () {
-    if (cartItems === null) return;
+        let numItems = cartItems.length;
+        sumTotal = 0;
+        for (let i = 0; i < cartItems.length; i++) {
+            sumTotal += (parseFloat(cartItems[i].Amount) * parseFloat(cartItems[i].Product_Price));
+           // console.log("Promise : ", cartItems[i]);
+        }
+    
+        document.getElementById("PCSumNumItem").innerHTML = numItems;
+        document.getElementById("PCPriceSum").innerHTML = sumTotal;
+    });
+    
 
-    let numItems = cartItems.length;
-
-    for (let i = 0; i < cartItems.length; i++) {
-        sumTotal += (cartItems[i].Amount * cartItems[i].Product_Price);
-    }
-
-    document.getElementById("PCSumNumItem").innerHTML = numItems;
-    document.getElementById("PCPriceSum").innerHTML = sumTotal;
 }
 
-let varifyAndProceed = function () {
+
+
+function varifyAndProceed() {
     if (user == null) {
         alert("Please sign in");
         window.location.href = "Login.html";
@@ -53,14 +109,23 @@ let varifyAndProceed = function () {
 
     //Upon clicking buy on the pop up
     document.getElementById("Buy-btn").addEventListener('click', function () {
-        proceedToBuy();
+         console.log('about to buy');
+
+         const promise = new Promise((resolve )=>{
+            resolve(proceedToBuy());
+         });
+         
+         promise.then(()=>{
+            DeleteItemFromCart()
+            promise.then(()=>{
+                setVariables();
+            });
+         })
     });
 
 }
 
-let proceedToBuy = function () {
-
-
+function proceedToBuy() {
     const buyer = user.UserID;
     let transDate = new Date();
     let dd = String(transDate.getDate()).padStart(2, '0');
@@ -69,61 +134,73 @@ let proceedToBuy = function () {
     transDate = mm + '/' + dd + '/' + yyyy;
 
     for (let i = 0; i < cartItems.length; i++) {
-
-        for (let j = 0; j < parseInt(cartItems[i].Amount); j++) {
-            cartItems[i].Quantity--;
+    
+        // Go to the next item if the quantity of the item in cart is 0
+        if(parseInt(cartItems[i].Current_Quantity) === 0){
+            alert(`${cartItems[i].Product_Name} is currently out of stock. \n You will be notified once we have stock`);
+            continue;
+        }
+        for (let j = 0; j < parseInt(cartItems[i].Amount); j++) {  
+            cartItems[i].Current_Quantity--;
             let prodID = cartItems[i].Product_ID;
             let balance = parseFloat(user.Balance) - parseFloat(cartItems[i].Product_Price);
             let Quant = parseInt(cartItems[i].Current_Quantity) - 1;
-
-
-
             console.log(transDate, prodID, buyer, balance, Quant);
-            $.getJSON(buyURL, {
-                ProductID: prodID,
-                Buyer: buyer,
-                TransDate: transDate,
-                Balance: balance,
-                Quantity: Quant
-            }, function (confirmation) {
-                console.log(confirmation);
-                if (confirmation === "1") {
-                    const updateUserURl = "https://lamp.ms.wits.ac.za/~s1814731/MPphpfiles/MPReturnUser.php";
-                    $.getJSON(updateUserURl, {
-                        username: user.UserName
-                    }, function (result) {
-                        if (result[0] !== '') {
-                            localStorage.removeItem('user');
-                            localStorage.setItem('user', JSON.stringify(result[0]));
-                            user = JSON.parse(localStorage.getItem("user"));
-                            console.log(JSON.parse(localStorage.getItem('user')));
-                            console.log("Product(s) successfully purchased");
 
-                            //Minuses one from the database
-                            $.getJSON(updateCartUrl, {
-                                userID: user.UserID,
-                                product_ID: cartItems[i].Product_ID
-                            }, function (update) {
-                                console.log(cartItems[i].Product_ID + ' minused')
-                            })
+            ProcessElement(prodID,buyer,transDate,transDate,balance,Quant);
+        }
+        
+    }
+    
+    
+    
 
-                        }
-                    });
+}
 
+function ProcessElement(prodID,buyer,transDate,transDate,balance,Quant){
+
+    $.getJSON(buyURL, {
+        ProductID: prodID,
+        Buyer: buyer,
+        TransDate: transDate,
+        Balance: balance,
+        Quantity: Quant
+    }, confirmation => {
+        console.log(confirmation);
+        if (confirmation === "1") {
+            const updateUserURl = "https://lamp.ms.wits.ac.za/~s1814731/MPphpfiles/MPReturnUser.php";
+            $.getJSON(updateUserURl, {
+                username: user.UserName
+            }, function (result) {
+                if (result[0] !== '') {
+                    localStorage.removeItem('user');
+                    localStorage.setItem('user', JSON.stringify(result[0]));
+                    user = JSON.parse(localStorage.getItem("user"));
+                    console.log(JSON.parse(localStorage.getItem('user')));
+                    console.log("Product successfully purchased");
+                     MinusFromDatabase(user.UserID,prodID);   
                 }
             });
 
         }
-
-        
-
-        //window.location.href = "Homepage.html";
-    }
-    DeleteItemFromCart();
+    });
 
 }
 
-let DeleteItemFromCart = function () {
+function MinusFromDatabase(uID , pID){
+ //Minuses one from the database
+ $.getJSON(updateCartUrl, {
+    userID: uID,
+    product_ID: pID
+}, function (update) {
+    console.log(pID + ' minused');
+    DeleteItemFromCart();
+
+});
+
+}
+
+function DeleteItemFromCart() {
     // Here we delete an 
     $.getJSON(cartUrl, {userID: user.UserID}, function (results) {
         results.map(function (item) {
@@ -135,15 +212,17 @@ let DeleteItemFromCart = function () {
                     product_ID: item.Product_ID
                 }, function (response) {
                     console.log(response);
+                    setVariables();
                 })
             }
 
         }).join('');
     });
 
-    setVariables();
+    
 }
+
 
 ConfirmPurchase.setAttribute("onclick", "varifyAndProceed()");
 
-setVariables();
+
